+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-8.2.0
+8.2.0 (2025-11-20)
 ------------------
 
 * Added `Securepay` to the `Processor` enum.